extract class for cli version extraction
